Product Controller Unit Tests
Produced unit Tests for the Product Controller. These cover the following:

• The correct Error is returned when attempting to update a product with the incorrect Id
• The correct Error is returned when attempting to pass an empty update
• A new product can be created using the controller

Extra Additions:
• Modified the function in the Home Controller Tests to populate the database

Important Notes:
Although all three Product Controller tests passed, when the product is created: it is not added to the database,
This is because the function requires an image file to proceed, even though no models have this as a required variable (using Code Annotations)
This appears to be mirrored in the front-end, although further Selenium tests would be required to confirm.