Added Product Controller: Delete Product Unit Test
Added a Fourth Unit Test to the Product Controller Unit Tests, which intended to test that created products can be deleted from the Database.

However, this test failed. This further highlights the oversight in the application, where functions using the product model assume an image is present, despite there being no validation or annotations that ensure its existence in a created product model.

This is intended to be fixed; either through adding validation to the Product Model or adjusting how this variable is handled in relevant functions