Added Shopping Cart Controller - Cart Total Unit Test
Added a unit test for the Shopping Cart Controller to check that the produced total matched the sum of all products in the cart.

This test failed. The problem has been sourced to the GetShoppingCartTotal() function in ShoppungCartModel.cs. Instead of using each products' price variable, the id is used: resulting in an incorrect total.