Selenium Tests for Product Display
Two Selenium Tests, each relating to the retrieval of Products on the main Index page.

ProductsDisplayedOnLoad() is used to check all products are returned upon Login.
To achieve this, IDs had to be added to the Home MVC. Test passed first try.

ProductDescriptionsDisplayed() is used to check the correct description is shown for each product.
Test Failed, 'Product.Desciption' was called as plaintext instead of a model reference. This was amended, and now test passes