Selenium Test to check Product Browsing
NavigateToProductBrowse() is used to test users can navigate to the Home Controller view used for displaying all products.

To achieve this test, IDs were added to _LoginPartial.cshtml

The Test failed. This is due to the Home controller returning the wrong Asp-Action. This was altered, so Products is now returned instead of Privacy. The test now passes