AdminProductManagement()
This Selenium test is used to validate that Admin-role users can access the product management page, and once there, create a new product.

To achieve this test, IDs were added, and a local image file was required. Therefore, with how selenium handles image directories, a new image path would be required on a different machine.

The Test passes